fix #7
