+++ src/app/(auth)/layout.tsx
@@ -1,0 +1,37 @@
+import "@/app/globals.css";
+import type { Metadata } from "next";
+import { Inter } from "next/font/google";
+import Image from "next/image";
+
+const inter = Inter({ subsets: ["latin"] });
+
+export const metadata: Metadata = {
+  title: "Psycho - الله معاك",
+  description: "i hope you die soon",
+};
+
+export default function RootLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return (
+    <html lang="en" suppressHydrationWarning>
+      <body className={inter.className}>
+        <div className="flex h-[100dvh]">
+          <div className="hidden w-2/3 bg-destructive md:block">
+            <Image
+              width={0}
+              height={0}
+              className="hidden h-screen w-full object-cover md:block"
+              src="AssetLogin.svg"
+              alt="hi mom"
+              loading="eager"
+            />
+          </div>
+          {children}
+        </div>
+      </body>
+    </html>
+  );
+}
